refactor: move all loggers to checker packages
