+++ packages/checker-eslint/package.json
@@ -42,12 +42,13 @@
     "tiny-invariant": "^1.3.1"
   },
   "peerDependencies": {
-    "vite-plugin-checker": ">= 0.7.0",
-    "eslint": "*"
+    "eslint": "*",
+    "vite-plugin-checker": ">= 0.7.0"
   },
   "devDependencies": {
-    "vite-plugin-checker": "workspace:*",
+    "@types/eslint": "^8.44.0",
     "eslint": "^8.11.0",
-    "tsup": "^6.7.0"
+    "tsup": "^6.7.0",
+    "vite-plugin-checker": "workspace:*"
   }
 }
